fix(rules): explicitly import rules to trigger side effects
This adds an explicit import for the rules index file, to trigger the registration side effects of each rule being imported. This already happened in the CLI due to the import being present in "autoload.ts", but was missing in programmatic usage. This resulted in "rules.loadAll" loading no rules.